Using of compound memory buffer for WebM read

+++ worker/include/MemoryBuffer.hpp
@@ -10,10 +10,10 @@ class MemoryBuffer
 {
 public:
     virtual ~MemoryBuffer() = default;
-    virtual size_t GetSize() const = 0;
+    virtual uint64_t GetSize() const = 0;
     virtual uint8_t* GetData() = 0;
     virtual const uint8_t* GetData() const = 0;
-    bool IsEmpty() const { return 0UL == GetSize() || nullptr == GetData() ; }
+    bool IsEmpty() const { return 0U == GetSize() || nullptr == GetData() ; }
 };
 
 } // namespace RTC
